FacePrivacy/PRO-Face/requirements.txt
absl-py==1.2.0
asttokens==2.0.5
backcall==0.2.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
decorator==5.1.1
dominate==2.7.0
executing==0.8.3
fonttools==4.34.4
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.47.0
idna==3.3
install==1.3.5
ipython==8.4.0
jedi==0.18.1
joblib==1.1.0
kiwisolver==1.4.4
kornia==0.6.7
lpips==0.1.4
Markdown==3.4.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
natsort==8.1.0
numpy==1.23.1
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.29
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-msssim==0.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
stack-data==0.2.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
threadpoolctl==3.1.0
torch==1.12.0+cu116
torchaudio==0.12.0+cu116
torchvision==0.13.0+cu116
tqdm==4.64.0
traitlets==5.2.2.post1
typing_extensions==4.3.0
urllib3==1.26.10
wcwidth==0.2.5
Werkzeug==2.1.2
